refactor ngff meta file
